+++ filenames.gypi
@@ -64,6 +64,7 @@
       'lib/renderer/inspector.js',
       'lib/renderer/override.js',
       'lib/renderer/security-warnings.js',
+      'lib/renderer/web-frame-init.js',
       'lib/renderer/window-setup.js',
       'lib/renderer/web-view/guest-view-internal.js',
       'lib/renderer/web-view/web-view.js',
