+++ lib/renderer/init.js
@@ -3,7 +3,6 @@
 const events = require('events')
 const path = require('path')
 const Module = require('module')
-const resolvePromise = Promise.resolve.bind(Promise)
 
 // We modified the original process.argv to let node.js load the
 // init.js, we need to restore it here.
@@ -26,7 +25,6 @@ var v8Util = process.atomBinding('v8_util')
 v8Util.setHiddenValue(global, 'ipc', new events.EventEmitter())
 
 // Use electron module after everything is ready.
-const electron = require('electron')
 
 const {
   warnAboutNodeWithRemoteContent,
@@ -40,40 +38,7 @@ const {
   shouldLogSecurityWarnings
 } = require('./security-warnings')
 
-// Call webFrame method.
-electron.ipcRenderer.on('ELECTRON_INTERNAL_RENDERER_WEB_FRAME_METHOD', (event, method, args) => {
-  electron.webFrame[method](...args)
-})
-
-electron.ipcRenderer.on('ELECTRON_INTERNAL_RENDERER_SYNC_WEB_FRAME_METHOD', (event, requestId, method, args) => {
-  const result = electron.webFrame[method](...args)
-  event.sender.send(`ELECTRON_INTERNAL_BROWSER_SYNC_WEB_FRAME_RESPONSE_${requestId}`, result)
-})
-
-electron.ipcRenderer.on('ELECTRON_INTERNAL_RENDERER_ASYNC_WEB_FRAME_METHOD', (event, requestId, method, args) => {
-  const responseCallback = function (result) {
-    resolvePromise(result)
-      .then((resolvedResult) => {
-        event.sender.send(`ELECTRON_INTERNAL_BROWSER_ASYNC_WEB_FRAME_RESPONSE_${requestId}`, null, resolvedResult)
-      })
-      .catch((resolvedError) => {
-        if (resolvedError instanceof Error) {
-          // Errors get lost, because: JSON.stringify(new Error('Message')) === {}
-          // Take the serializable properties and construct a generic object
-          resolvedError = {
-            message: resolvedError.message,
-            stack: resolvedError.stack,
-            name: resolvedError.name,
-            __ELECTRON_SERIALIZED_ERROR__: true
-          }
-        }
-
-        event.sender.send(`ELECTRON_INTERNAL_BROWSER_ASYNC_WEB_FRAME_RESPONSE_${requestId}`, resolvedError)
-      })
-  }
-  args.push(responseCallback)
-  electron.webFrame[method](...args)
-})
+require('./web-frame-init')()
 
 // Process command line arguments.
 let nodeIntegration = 'false'
