+++ lib/sandboxed_renderer/init.js
@@ -32,6 +32,8 @@ const preloadModules = new Map([
   ['timers', require('timers')]
 ])
 
+require('../renderer/web-frame-init')()
+
 // Pass different process object to the preload script(which should not have
 // access to things like `process.atomBinding`).
 const preloadProcess = new events.EventEmitter()
